requirements.txt
groq>=1.1.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
typer[all]>=0.12.0
rich>=13.7.0
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.31.0
pydantic>=2.0.0
